+++ lab/exp-bib2bibentry.R
@@ -4,3 +4,8 @@
 ################################################################################
 
 library(biblio)
+
+Bib <- read_bib(x = file.path(path.package("biblio"),
+        "LuebertPliscoff.bib"))
+
+bib2bibentry(Bib[1:5, ])