+++ NEWS.md
@@ -5,14 +5,14 @@
 
 - New function `bib2bibentry()` for converting `lib_df` objects into
   `bibentry` objects.
+- Coerce methods implemented for `lib_df` objects.
 
 ### Improvements
 
 - Improved function `read_bib()` using regular expressions.
 
 
-biblio 0.0.7
-============
+# biblio 0.0.7
 
 ### New Features